🔧 Add .DS_Store and workspace.xml files to project 🔧 Add settings.json file for VS Code configuration ♻️ Refactor RabbitMQConfig to use constant for queue name ♻️ Refactor MovementController to add registerMovement endpoint ✨ Add ReportController with generate and get endpoints ✨ Add CustomerFilterDTO for filtering customers ♻️ Refactor Account entity to include customerId and movementList
✨ (Movement.java): Add Serializable interface to Movement class for serialization support
✨ (Movement.java): Change date field type from LocalDateTime to Date for compatibility
✨ (Movement.java): Rename balance field to initialBalance for clarity
✨ (Movement.java): Add relationship annotations for Account entity in Movement class
✨ (Report.java): Create Report entity with fields for customerName, availableBalance, and relationships with Account and Movement entities
✨ (ResourceNotFoundException.java): Create custom exception class ResourceNotFoundException with HttpStatus.NOT_FOUND
✨ (ReportListener.java): Create ReportListener class to listen for customer responses and generate reports based on movements
✨ (AccountRepository.java): Add method findByCustomerId to AccountRepository interface
✨ (MovementRepository.java): Add method findByAccountAndDateBetween to MovementRepository interface
✨ (ReportRepository.java): Create ReportRepository interface with method findByAccount_AccountNumber
✨ (AccountService.java): Add exception handling for getAllAccounts, getAccountById, saveAccount, and deleteAccount methods

📝 (MovementService.java): Add support for handling exceptions and improve error messages in getAllMovements, getMovementById, saveMovement, deleteMovement, and registerMovement methods
📝 (ReportService.java): Refactor generateReport method to use CustomerFilterDTO and improve error handling in getReportsByAccountNumber method
🔧 (application.properties): Update RabbitMQ configuration with default values for host, port, username, and password
🔧 (RabbitMQConfig.java): Refactor RabbitMQ configuration to use properties for queue, exchange, and routing key names
🔧 (CustomerController.java): Update method parameter name from customerId to id in updateCustomer method

✨ (ReportDTO.java): Add ReportDTO class to handle report data transfer object
♻️ (Customer.java): Refactor customerId field to id in Customer entity for better naming convention
📝 (ResourceNotFoundException.java): Add ResourceNotFoundException class to handle resource not found exceptions
✨ (CustomerListener.java): Add CustomerListener class to listen for customer requests and send reports
🔧 (application.properties): Update application properties to include RabbitMQ configurations and server port
🔧 (docker-compose.yml): Update docker-compose file to include RabbitMQ username and password configurations

+++ account-movements-microservice/src/main/java/com/alargo/account_movements_microservice/services/AccountService.java
@@ -1,6 +1,8 @@
 package com.alargo.account_movements_microservice.services;
 
 import com.alargo.account_movements_microservice.entity.Account;
+import com.alargo.account_movements_microservice.exception.CustomException;
+import com.alargo.account_movements_microservice.exception.ResourceNotFoundException;
 import com.alargo.account_movements_microservice.repository.AccountRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -17,21 +19,56 @@ public class AccountService {
         this.accountRepository = accountRepository;
     }
 
+
     public List<Account> getAllAccounts() {
-        return accountRepository.findAll();
+        try {
+            if (accountRepository.findAll().isEmpty()) {
+                throw new ResourceNotFoundException("No hay cuentas registradas");
+            }
+
+            return accountRepository.findAll();
+        } catch (ResourceNotFoundException e) {
+            throw e;
+        } catch (Exception e) {
+            throw new CustomException("Error al obtener todas las cuentas: " + e.getMessage());
+        }
     }
 
     public Account getAccountById(Long id) {
-        return accountRepository.findById(id).orElse(null);
+        try {
+            if (!accountRepository.existsById(id)) {
+                throw new ResourceNotFoundException("Cuenta no encontrada");
+            }
+            return accountRepository.findById(id)
+                    .orElseThrow(() -> new ResourceNotFoundException("Cuenta no encontrada"));
+        } catch (ResourceNotFoundException e) {
+            throw e;
+        } catch (Exception e) {
+            throw new CustomException("Error al obtener la cuenta por ID: " + e.getMessage());
+        }
     }
 
     public Account saveAccount(Account account) {
-        return accountRepository.save(account);
+        try {
+            return accountRepository.save(account);
+        } catch (Exception e) {
+            throw new CustomException("Error al guardar la cuenta: " + e.getMessage());
+        }
     }
 
     public void deleteAccount(Long id) {
-        accountRepository.deleteById(id);
+        try {
+            if (!accountRepository.existsById(id)) {
+                throw new ResourceNotFoundException("Cuenta no encontrada");
+            }
+            accountRepository.deleteById(id);
+        } catch (ResourceNotFoundException e) {
+            throw e;
+        } catch (Exception e) {
+            throw new CustomException("Error al eliminar la cuenta: " + e.getMessage());
+        }
     }
+
 }
